chore: move draw_felt and draw_word functions to rng.rs
